adding wait functions and test duration timer.

+++ GitHubProjectFolder/config/AutomationConfig.cs
@@ -8,8 +8,8 @@ namespace GitHub.config
 {
     public class AutomationConfig
     {
-        public const int DefaultTimeout = 30;
-        public const int DefaultTimeoutForLongWaits = 60;
+        public const double DefaultTimeout = 30;
+        public const double DefaultTimeoutForLongWaits = 60;
         public static IWebDriver Driver = new ChromeDriver();
         public WebDriverWait Wait = new WebDriverWait(Driver, TimeSpan.FromSeconds(DefaultTimeout));
